==> 02_kidney_data_GSE129798.Rmd <==
---
title: "02_kidney_data_GSE129798"
author: "YinCY"
date: "2023-03-01"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# prepare dev cell data
data from GSE129798
```{r}
library(DropletUtils)
library(scran)
library(scuttle)
library(scater)
library(magrittr)

dev_cell_sce <- read10xCounts(samples = "../data/kidney/kidney-map/GSE129798/Mouse_Adult_DGE/", 
                              type = "sparse", 
                              col.names = TRUE)
dev_cell_sce %>% colData()
dev_cell_sce %>% rowData() %>% .$ID %>% duplicated() %>% table

library(AnnotationHub)
ah <- AnnotationHub()
query(ah, c("EnsDb", "Mus Musculus"))
ens107 <- ah[["AH104895"]]

rowData(dev_cell_sce)$ensembl <- mapIds(ens107, 
                                        keys = rowData(dev_cell_sce)$ID, 
                                        keytype = "SYMBOL", 
                                        column = "GENEID")

rowData(dev_cell_sce)$ensembl %>% is.na() %>% table()
dev_cell_sce <- dev_cell_sce[!is.na(rowData(dev_cell_sce)$ensembl), ]
rownames(dev_cell_sce) <- rowData(dev_cell_sce)$ensembl
rowData(dev_cell_sce)$symbol <- rowData(dev_cell_sce)$ID
rowData(dev_cell_sce)$ID <- NULL
rowData(dev_cell_sce)

colData(dev_cell_sce)$cell_ids <- colData(dev_cell_sce)$Barcode
colData(dev_cell_sce)$Barcode <- NULL

colData(dev_cell_sce)$location <- substr(colData(dev_cell_sce)$cell_ids, 1, 2)
colData(dev_cell_sce)$location <- sapply(colData(dev_cell_sce)$location, function(x){
  switch(x, 
         "Z1" = "Cortex", 
         "Z2" = "outer medulla", 
         "Z3" = "inner medulla")
})
colData(dev_cell_sce)$batch <- substr(colData(dev_cell_sce)$cell_ids, 1, 4)
colData(dev_cell_sce)$gender <- substr(colData(dev_cell_sce)$cell_ids, 4, 4)
colData(dev_cell_sce)

# retain only the male data
dev_cell_sce <- dev_cell_sce[, dev_cell_sce$gender == "M"]
table(colData(dev_cell_sce)$gender);table(colData(dev_cell_sce)$batch);table(colData(dev_cell_sce)$location)
```


```{r}
# quality control
is_mito <- grepl("^mt-", rowData(dev_cell_sce)$symbol, ignore.case = F)
is_mito %>% table
rowData(dev_cell_sce)$symbol[is_mito] %>% sort
colData(dev_cell_sce)$mito_percent <- round((colSums2(assay(dev_cell_sce[is_mito, ])) / colSums2(assay(dev_cell_sce))) * 100, 4)

stats <- perCellQCMetrics(dev_cell_sce, subsets = list(mito = is_mito))
qc <- perCellQCFilters(stats, sub.fields = "subsets_mito_percent")
qc %>% as.matrix() %>% colSums()

dev_cell_sce %>% colData() %>% as.data.frame() %>% cbind(as.data.frame(qc)) %>% 
  ggplot(aes(location, mito_percent)) +
  geom_jitter(aes(color = low_n_features), alpha = 1/3)

dev_cell_sce %>% colData() %>% as.data.frame() %>% cbind(as.data.frame(qc)) %>% 
  ggplot(aes(location, mito_percent)) +
  geom_jitter(aes(color = high_subsets_mito_percent), alpha = 1/3)

# filter based on number of features
dev_cell_sce <- dev_cell_sce[, !qc$low_n_features]
```

```{r}
# normalization
set.seed(101)
qclusters <- quickCluster(dev_cell_sce, 
                         min.size = 50, 
                         method = "igraph", 
                         use.ranks = T, 
                         d = 50)
qclusters %>% table
colData(dev_cell_sce)$qclusters <- qclusters

sf <- pooledSizeFactors(dev_cell_sce, clusters = colData(dev_cell_sce)$qclusters)
sizeFactors(dev_cell_sce) <- sf
colData(dev_cell_sce)

dev_cell_sce <- logNormCounts(dev_cell_sce)
dev_cell_sce
```


```{r}
# feature selection
library(stringr)
library(BiocParallel)

dec_dev_cell <- modelGeneVar(dev_cell_sce, block = colData(dev_cell_sce)$batch)
dec_dev_cell[order(dec_dev_cell$bio, decreasing = T), ]
```

```{r}
par(mfrow = c(1, 3))
for(i in 1:ncol(dec_dev_cell$per.block)){
    plot(metadata(dec_dev_cell$per.block[[i]])$mean, 
         metadata(dec_dev_cell$per.block[[i]])$var, 
         pch = 19, cex = 0.5, 
         main = paste0("model_", names(dec_dev_cell$per.block)[[i]]))
    curve(metadata(dec_dev_cell$per.block[[i]])$trend(x), col = "red", lwd = 2, add = TRUE)
}
```



```{r}
dev_cell_hvgs <- getTopHVGs(stats = dec_dev_cell, 
                            var.field = "bio",
                            var.threshold = 0, 
                            fdr.threshold = 0.01)
dev_cell_hvgs %>% str

# choosing the number of PCs
set.seed(102)
npcs <- calculatePCA(dev_cell_sce, subset_row = dev_cell_hvgs, ncomponents = 50)
percentvar <- attr(npcs, which = "percentVar")

library(PCAtools)
elbow <- findElbowPoint(percentvar)

df <- data.frame(PCs = 1:dim(npcs)[2], 
                 percentvar = percentvar)

df %>% 
    ggplot(aes(PCs, percentvar)) + 
    geom_line() + 
    geom_point() +
    geom_vline(xintercept = elbow) +
    scale_y_continuous(trans = "log2")

reducedDim(dev_cell_sce, "PCA") <- npcs[, 1:14]

set.seed(103)
dev_cell_sce <- runTSNE(x = dev_cell_sce, 
                        dimred = "PCA", 
                        perplexity = 10, 
                        BPPARAM = MulticoreParam(workers = 12))

plotTSNE(dev_cell_sce, 
         colour_by = "qclusters", 
         point_size = 0.5, 
         text_by = "qclusters") +
  theme(legend.position = "none")
```


```{r}
# graph based clustering
library(bluster)
library(patchwork)

set.seed(104)
clusters <- clusterRows(reducedDim(dev_cell_sce, "PCA"), 
                        BLUSPARAM = NNGraphParam(k = 10, 
                                                 cluster.fun = "walktrap"))
clusters %>% table
colData(dev_cell_sce)$clusters <- factor(clusters)
colData(dev_cell_sce)

source("functions/idconv.R")
plotTSNE(dev_cell_sce, 
         # colour_by = idconv("Corin"),
         colour_by = "location",
         point_size = 0.5, 
         text_by = "clusters") +
  theme(legend.position = "none") +
plotTSNE(dev_cell_sce, 
         colour_by = idconv("Cdh13"),
         # colour_by = "clusters",
         point_size = 0.5, 
         text_by = "clusters") +
  theme(legend.position = "none") 
plotTSNE(dev_cell_sce, 
         colour_by = idconv("Pxylp1"),
         # colour_by = "clusters",
         point_size = 0.5, 
         text_by = "clusters") +
  theme(legend.position = "none")
```


```{r}
# doublet clusters detection
library(scDblFinder)
set.seed(105)
dbs <- findDoubletClusters(dev_cell_sce, clusters = colData(dev_cell_sce)$clusters)
dbs %>% as.data.frame() %>% dplyr::arrange(median.de)
# doublets <- c(25, 30, 31, 32, 33, 35, 36, 38, 39, 40, 41)
```



# sub-clusters
```{r}
# 1 MTAL CTAL
c1 <- dev_cell_sce[, dev_cell_sce$clusters == 1]

hvgs_mtal <- readxl::read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                                sheet = "MTAL_1020") %>% 
    dplyr::filter(FDR < 0.05) %>% 
    dplyr::pull(Gene_id) 

hvgs_ctal <- readxl::read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                                sheet = "CTAL_1092") %>% 
    dplyr::filter(FDR < 0.05) %>% 
    dplyr::pull(Gene_id)

hvgs <- c(hvgs_ctal, hvgs_mtal) %>% unique()
hvgs <- hvgs[hvgs %in% rownames(c1)]

set.seed(101)
c1 <- runPCA(c1, subset_row = hvgs)

set.seed(102)
c1 <- runTSNE(c1, dimred = "PCA", perplexity = 50)

set.seed(103)
clusters_c1 <- clusterRows(reducedDim(c1, "PCA"), 
                           BLUSPARAM = NNGraphParam(k = 20))
clusters_c1 %>% table
colLabels(c1) <- clusters_c1
# 1.1 LOH_CTAL, 1.2 LOH_MTAL
colLabels(c1) <- paste("1.", colLabels(c1), sep = "")
labels <- as.character(dev_cell_sce$clusters)
labels[labels == "1"] <- colLabels(c1)
dev_cell_sce$clusters <- labels

plotTSNE(c1, 
         # colour_by = idconv("Shd"),
         colour_by = "label")
```


```{r}
# 8 VSMC, Pericytes
c8 <- dev_cell_sce[, dev_cell_sce$clusters == 8]
hvgs <- c("ENSMUSG00000024620", "ENSMUSG00000032911", "ENSMUSG00000001349")

set.seed(106)
c8 <- runPCA(c8, exprs_values = "logcounts", 
             ncomponents = 2, 
             subset_row = hvgs)

set.seed(107)
c8 <- runTSNE(c8, perplexity = 40, dimred = "PCA")

set.seed(108)
clusters_c8 <- clusterRows(reducedDim(c8, "PCA"), 
                           BLUSPARAM = NNGraphParam(k = 20))
clusters_c8 %>% table
colLabels(c8) <- clusters_c8

colLabels(c8) <- sapply(clusters_c8, FUN = function(x){
  switch(as.character(x), 
         "1" = "2",
         "2" = "1", 
         "3" = "1", 
         "4" = "2", 
         "5" = "2", 
         "6" = "2")
})
# 8.1 VSMC 8.2 Pericytes
colLabels(c8) <- paste("8.", colLabels(c8), sep = "")
labels <- as.character(dev_cell_sce$clusters)
labels[labels == "8"] <- colLabels(c8)
dev_cell_sce$clusters <- labels

source("functions/idconv.R")
plotReducedDim(c8, dimred = "TSNE", 
               colour_by = idconv("Cnn1"), 
               text_by = "label")
```


```{r}
# 7 ATL, DTL3
c7 <- dev_cell_sce[, grep("^7", dev_cell_sce$clusters)]
idconv("Osr2")
hvgs <- c("ENSMUSG00000004655", "ENSMUSG00000043461", "ENSMUSG00000022330")
dev_cell_hvgs <- getTopHVGs(stats = dec_dev_cell, 
                            var.field = "bio",
                            var.threshold = 0, 
                            fdr.threshold = 0.01)
dev_cell_hvgs %>% str


set.seed(106)
c7 <- runPCA(c7, exprs_values = "logcounts", 
             ncomponents = 10, 
             subset_row = hvgs)

set.seed(107)
c7 <- runTSNE(c7, perplexity = 30, dimred = "PCA")

set.seed(108)
clusters_c7 <- clusterRows(reducedDim(c7, "PCA"), 
                           BLUSPARAM = NNGraphParam(k = 30))
clusters_c7 %>% table
colLabels(c7) <- clusters_c7

# 1: ATL
# 2: DTL3
colLabels(c7) <- sapply(clusters_c7, FUN = function(x){
  switch(as.character(x), 
         "1" = "1",
         "2" = "2", 
         "3" = "1", 
         "4" = "1", 
         "5" = "1", 
         "6" = "1", 
         "7" = "1", 
         "8" = "2", 
         "9" = "1")
})

colLabels(c7) <- paste("7.", colLabels(c7), sep = "")
labels <- as.character(dev_cell_sce$clusters)
labels[grep("^7", labels)] <- colLabels(c7)
dev_cell_sce$clusters <- labels


plotTSNE(c7, 
         colour_by = "label",
         # colour_by = idconv("Sptssb"),
         text_by = "label")
```


```{r}
# 12 Cd4T, Cd8T,NK
c12 <- dev_cell_sce[, dev_cell_sce$clusters == 12]
hvgs <- c("ENSMUSG00000032094", "ENSMUSG00000002033", 
          "ENSMUSG00000030579", "ENSMUSG00000023274", 
          "ENSMUSG00000053977", "ENSMUSG00000053044")

set.seed(101)
c12 <- runPCA(c12, 
              subset_row = hvgs, 
              ncomponents = 4)

set.seed(102)
c12 <- runTSNE(c12, dimred = "PCA", preplexity = 30)

set.seed(103)
c12_clusters <- clusterRows(reducedDim(c12, "PCA"), 
                            BLUSPARAM = NNGraphParam(k = 50))
c12_clusters %>% table
colLabels(c12) <- c12_clusters

colLabels(c12) <- sapply(c12_clusters, FUN = function(x){
  switch(as.character(x), 
         "1" = "1",
         "2" = "3", 
         "3" = "1", 
         "4" = "1", 
         "5" = "2", 
         "6" = "1")
})
# 1,3,4,6: Cd4 T cells
# 5: Cd8 T cells
# 2: NK cells

colLabels(c12) <- paste("12.",colLabels(c12), sep = "") 

labels <- as.character(dev_cell_sce$clusters)
labels[labels == "12"] <- colLabels(c12)
dev_cell_sce$clusters <- labels

library(patchwork)
plotTSNE(c12, 
         colour_by = idconv("Cd8a"),
         # colour_by = "label",
         text_by = "label") +
plotTSNE(c12, 
         colour_by = idconv("Cd4"),
         # colour_by = "label",
         text_by = "label")
```


```{r}
# 17, 29: CCD_IC, OMCD_IC, IMCD_IC
cs <- dev_cell_sce[, dev_cell_sce$clusters %in% c(17, 29)]
hvgs_ccd <- readxl::read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                               sheet = "CCD_797") %>% 
    dplyr::filter(FDR < 0.05) %>% 
    dplyr::pull(Gene_id)

hvgs_omcd <- readxl::read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                                sheet = "OMCD_757") %>% 
    dplyr::filter(FDR < 0.05) %>% 
    dplyr::pull(Gene_id)

hvgs_imcd <- readxl::read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                                sheet = "IMCD_1951") %>% 
    dplyr::filter(FDR < 0.05) %>% 
    dplyr::pull(Gene_id)

hvgs <- c(hvgs_ccd, hvgs_imcd, hvgs_omcd) %>% unique()
hvgs <- hvgs[hvgs %in% rownames(cs)]

set.seed(101)
cs <- runPCA(cs, exprs_values = "logcounts", 
             ncomponents = 10, 
             subset_row = hvgs)

set.seed(102)
cs <- runTSNE(cs, perplexity = 30, dimred = "PCA")

set.seed(103)
clusters_cs <- clusterRows(reducedDim(cs, "PCA"), 
                           BLUSPARAM = NNGraphParam(k = 35))
clusters_cs %>% table
colLabels(cs) <- clusters_cs

# 1: IMCD_IC, 2: CCD_IC, 3: OMCD_IC
colLabels(cs) <- paste("17_29.", colLabels(cs), sep = "") 
labels <- as.character(dev_cell_sce$clusters)
labels[labels %in% c(17, 29)] <- colLabels(cs)
dev_cell_sce$clusters <- labels


plotTSNE(cs, 
         # colour_by = "label", 
         colour_by = idconv("Fxyd2"),
         text_by = "label")
```


```{r}
# cell type annotation base on marker genes
markers <- list(
    endothelia = c("Cdh5", "Pecam1", "Kdr", "Nrp1"), 
    macrophage = c("C1qa", "C1qb", "C1qc", "Cx3cr1", "Lyz2"), 
    T_cell = c("Thy1", "Cd3g", "Cd3d"), 
    B_cell = c("Cd79a", "Cd79b"),
    NK = c("Tyrobp"), 
    DC = c("H2-Eb1", "H2-Ab1"),
    Monocytes = c("Csf1r"),
    VSMC = c("Cnn1"), 
    pericyte = c("Cspg4"), 
    Fibroblast = c("Dcn"), 
    PTS1 = c("Slc5a2", "Slc5a12", "Slc7a7", "Spp2", "Slc6a19", "Slc22a29"),
    PTS2 = c("Fxyd2", "Hrsp12"), 
    PTS3 = c("Atp11a", "Slc13a3", "Slc27a2"),
    CD_IC = c("Atp6v1g3", "Slc4a1", "Aqp6", "Slc26a4", "Hmx2"),
    CD_PC = c("Hsd11b2", "Aqp4", "Aqp2"),
    DCT = c("Pvalb", "Slc12a3"),
    LOH = c("Clcnka", "Slc12a1", "Umod", "Sptssb"), 
    LOH_DTL1 = c("Lbp"), 
    LOH_DTL2 = c("Aqp1", "Slc39a8", "Tspan6", "Nr2e3"), 
    LOH_DTL3 = c("Aqp1, Osr2"), 
    LOH_ATL = c("Sptssb"),
    LOH_MTAL = c("Clcnka", "Umod", "Slc12a1", "Pcsk6", "Tmem207", "Kcnt1"),
    LOH_CTAL = c("Shd", "Dusp9", "Rab6b", "Rps6ka6", "Fgf9"),
    CNT = c("Kcne1", "S100g"), 
    CCD = c("Calb1", "Slc4a1", "Pth1r", "Hmx2", "Insrr"),
    OMCD = c("Aqp2", "Aqp4", "Hsd11b2", "Slc4a1", "Slc26a7"), 
    IMCD = c("Wnt4", "Aqp2", "Aqp4")
)


cell_annotation <- c(
    "1.1" = "LOH_CTAL",
    "1.2" = "LOH_MTAL",
    "2" = "IMCD_PC", 
    "3" = "LOH_DTL2",
    "4" = "Monocytes",
    "5" = "Macrophages",
    "6" = "PTS2", 
    "7.1" = "LOH_ATL",
    "7.2" = "LOH_DTL3",
    "8.1" = "VSMC",
    "8.2" = "Pericytes",
    "9" = "Endothelial",
    "10" = "PTS3", 
    "11" = "Dendritic Cells", 
    "12.1" = "Cd4 T Cells", 
    "12.2" = "Cd8 T Cells",
    "12.3" = "NK Cells",
    "13" = "DCT", 
    "14" = "OMCD_PC",
    "15" = "Monocytes", 
    "16" = "Endothelial",
    "17_29.1" = "IMCD_IC",
    "17_29.2" = "CCD_IC",
    "17_29.3" = "OMCD_IC",
    "18" = "PTS1",
    "19" = "IMCD_PC", 
    "20" = "Fibroblast",
    "21" = "PTS1", 
    "22" = "CCD_PC", 
    "23" = "LOH_DTL1", 
    "24" = "B Cells", 
    "25" = "doublets", 
    "26" = "DCT", 
    "27" = "Endothelial", 
    "28" = "DCT", 
    "30" = "doublets",
    "31" = "doublets", 
    "32" = "doublets", 
    "33" = "doublets", 
    "34" = "CNT", 
    "35" = "doublets", 
    "36" = "doublets", 
    "37" = "Monocytes", 
    "38" = "doublets", 
    "39" = "doublets", 
    "40" = "doublets", 
    "41" = "doublets"
)
```


```{r}
# highlight each cluster and sub-cluster
for( n in unique(colData(dev_cell_sce)$clusters)){
    my_color <- rep("grey", length(colData(dev_cell_sce)$clusters))
    names(my_color) <- as.character(colData(dev_cell_sce)$clusters)
    my_color[colData(dev_cell_sce)$clusters == n] <- "red"
    
    plotReducedDim(object = dev_cell_sce, 
                   dimred = "TSNE", 
                   colour_by = "clusters", 
                   text_by = "clusters", 
                   point_size = 0.2) +
        scale_color_manual(values = my_color) +
        guides(colour = guide_legend(override.aes = list(size = 2)))
    
    ggsave(paste0("dev_image/the-", n, "th-cluster.pdf"))
}
```



```{r}
colData(dev_cell_sce)$cell_type <- cell_annotation[dev_cell_sce$clusters]
dev_cell_sce %>% colData()
```

```{r}
# remove doublets
dev_cell_sce <- dev_cell_sce[, dev_cell_sce$cell_type != "doublets"]
dev_cell_sce$cell_type %>% table(useNA = "ifany")
dev_cell_sce %>% colData

saveRDS(dev_cell_sce, "../data/kidney/kidney-map/GSE129798/dev_cell_sce.rds")
```



## annotate region information
```{r}
library(SingleR)
library(magrittr)

dev_cell_sce <- readRDS("../data/kidney/kidney-map/GSE150338/dev_cell_sce.rds")
ref_mouse_tubule_sce <- readRDS("../data/kidney/kidney-map/GSE150338/ref_mouse_tubule_sce.rds")
colData(ref_mouse_tubule_sce)

dev_cell_agg <- aggregateAcrossCells(dev_cell_sce, 
                                     ids = colData(dev_cell_sce)$clusters, 
                                     coldata.merge = list(mito_percent = median), 
                                     use.assay.type = "logcounts")

# cell_to_annotate <- c("PTS1", "PTS2", "PTS3", "LOH", "CD_PC", "CD_IC", "DCT")
# subs_dev_cell_tubule <- dev_cell_sce[, dev_cell_sce$cell_type %in% cell_to_annotate]

set.seed(109)
tubule_pred <- SingleR(test = dev_cell_agg, 
                       ref = ref_mouse_tubule_sce, 
                       labels = ref_mouse_tubule_sce$cell_type, 
                       assay.type.test = "logcounts", 
                       assay.type.ref = "tpm", 
                       de.method = "classic")

tubule_pred$labels %>% table
cs <- rownames(tubule_pred)
```

```{r}
# regional information
tubule_region <- data.frame(
    short_name = c("PTS1", "PTS2", "PTS3",
                   "DTL1", "DTL2", "DTL3",
                   "ATL", "MTAL", "CTAL",
                   "DCT", "CNT", "CCD",
                   "OMCD", "IMCD"), 
    log_name = c("initial segment of the proximal convoluted tubule",
                 "proximal straight tubule in cortical medullary lays", 
                 "last segment of the proximal straight tubule in outer stripe of outer medulla", 
                 
                 "short descending limb of the loop of Henle", 
                 "long descending limb of the loop of Henle in the outer medulla", 
                 "long descending limb of the loop of Henle in the inner medulla", 
                 
                 "thin ascending limb of the loop of Henle", 
                 "medullary thick ascending limb of the loop of Henle", 
                 "cortical thick ascending limb of the loop of Henle", 
                 
                 "distal convoluted tubule", 
                 "connecting tubule", 
                 "cortical collecting duct", 
                 
                 "outer medullary collecting duct", 
                 "inner medullary collecting duct"), 
    location = c("Cortex", "Cortex", "outer medulla", 
                 "out medulla", "outer medulla", "inner medulla", 
                 "inner medulla", "outer medulla", "Cortex/outer medulla", 
                 "Cortex", "Cortex", "Cortex", 
                 "outer medulla", "inner medulla")
)
```

```{r}
knitr::include_graphics("../data/kidney/kidney-map/GSE150338/tubule_region.png")
```



```{r}
# quality check
pdf("tubule_pre.pdf")
tubule_pred$scores %>% as.data.frame(row.names = cs) %>% as.matrix() %>% 
    pheatmap::pheatmap(angle_col = 315, 
                       cluster_cols = T, 
                       show_rownames = T,
                       show_colnames = T,
                       color = viridis::viridis(n = 1000, option = "H"), 
                       fontsize_row = 5, 
                       display_numbers = TRUE, 
                       fontsize_number = 4, 
                       number_format = "%.3f")
dev.off()

plotDeltaDistribution(tubule_pred, 
                      show = "delta.med")

marker_low_assignments <- pruneScores(results = tubule_pred, 
                                      nmads = 3)
marker_low_assignments %>% table
```


```{r}
colData(subs_dev_cell_tubule)$pre_segments <- tubule_pred$labels
colData(subs_dev_cell_tubule)$assignment_quality_score <- ifelse(pruneScores(tubule_pred) == FALSE, "high", "low")

cell_location <- setNames(tubule_region$location, tubule_region$short_name)
colData(subs_dev_cell_tubule)$pre_location <- cell_location[subs_dev_cell_tubule$pre_segments]

colData(subs_dev_cell_tubule) %>% as.data.frame() %>% dplyr::filter(assignment_quality_score == "high")

colData(subs_dev_cell_tubule)


# assignment quality
table(factor(subs_dev_cell_tubule$location, levels = c("Cortex", "outer medulla", "inner medulla")), 
      factor(subs_dev_cell_tubule$pre_segments, levels = c("PTS1", "PTS2", "PTS3", "DTL1",
                                                           "DTL2", "DTL3", "ATL", "MTAL", 
                                                           "CTAL", "DCT", "CNT", "CCD",
                                                           "OMCD", "IMCD"))) %>% 
  prop.table(margin = 2) %>% round(digits = 3) %>% knitr::kable()
```


```{r}
# remove low assignmented cells
subs_dev_cell_tubule <- subs_dev_cell_tubule[, subs_dev_cell_tubule$assignment_quality_score == "high"]

# for PTS1, PTS2 and PTS3, using pre_segments column to replace cell_type column
conditions <- subs_dev_cell_tubule$pre_segments %in% c("PTS1", "PTS2", "PTS3")
subs_dev_cell_tubule[, conditions]$cell_type <- subs_dev_cell_tubule[, conditions]$pre_segments

# for LOH, there 6 segments, DTL1, DTL2, DTL3, ATL, MTAL and CTAL
# the name will be named as LOH_segment
conditions <- subs_dev_cell_tubule$pre_segments %in% c("DTL1", "DTL2", "DTL3", "ATL", "MTAL", "CTAL")
subs_dev_cell_tubule[, conditions]$cell_type <- paste("LOH_", subs_dev_cell_tubule[, conditions]$pre_segments, sep = "")

# for distal convoluted tubule, there two segments DCT, CNT
conditions <- subs_dev_cell_tubule$pre_segments %in% c("DCT", "CNT")
subs_dev_cell_tubule[, conditions]$cell_type <- subs_dev_cell_tubule[, conditions]$pre_segments

# for collecting ducts, there three segments, CCD, OMCD, IMCD
# collecting ducts contains principal cells and intercalated cells, this information will be retained.
library(stringr)
library(stringr.plus)
conditions <- subs_dev_cell_tubule$pre_segments %in% c("CCD", "OMCD", "IMCD")
subs_dev_cell_tubule[, conditions]$cell_type <- ifelse(str_detect(subs_dev_cell_tubule[, conditions]$cell_type, "_"),  
                                                       paste(subs_dev_cell_tubule[, conditions]$pre_segments, str_extract_after(subs_dev_cell_tubule[, conditions]$cell_type, "_"), sep = "_"), 
                                                       subs_dev_cell_tubule[, conditions]$cell_type)

# remove unannotated LOH
subs_dev_cell_tubule <- subs_dev_cell_tubule[, subs_dev_cell_tubule$cell_type != "LOH"]

subs_dev_cell_tubule %>% 
  colData() %>% 
  as.data.frame() %>% 
  dplyr::select(cell_type, pre_segments) %>% 
  table
```

```{r}
library(patchwork)
plotTSNE(subs_dev_cell_tubule, colour_by = "cell_type") +
plotTSNE(subs_dev_cell_tubule, colour_by = "location")
```


```{r}
# assign the location information back to the dev_cell_sce object
others <- dev_cell_sce[, !(colData(dev_cell_sce)$cell_type %in% cell_to_annotate)] 
others %>% colData() %>% colnames(); subs_dev_cell_tubule %>% colData %>% colnames()
colData(others)$pre_segments <- NA
colData(others)$assignment_quality_score <- NA
colData(others)$pre_location <- NA
dev_cell_sce2 <- cbind(others, subs_dev_cell_tubule)

dev_cell_sce2 %>% colData() 

library(scater)
library(viridis)
plotReducedDim(dev_cell_sce2, 
               dimred = "TSNE", 
               colour_by = "cell_type", 
               text_by = "cell_type", 
               text_size = 2.5, 
               point_size = 0.5) +
  scale_color_viridis(discrete = TRUE, option = "D") +
  theme(legend.position = "none")

saveRDS(dev_cell_sce2, "../data/kidney/kidney-map/GSE129798/dev_cell_sce.rds")
```


```{r}
for( n in unique(colData(dev_cell_sce2)$cell_type)){
    my_color <- rep("grey", length(colData(dev_cell_sce2)$cell_type))
    names(my_color) <- as.character(colData(dev_cell_sce2)$cell_type)
    my_color[colData(dev_cell_sce2)$cell_type == n] <- "red"
    
    plotReducedDim(object = dev_cell_sce2, 
                   dimred = "TSNE", 
                   colour_by = "cell_type", 
                   text_by = "cell_type", 
                   point_size = 0.2, 
                   text_size = 2.5) +
        scale_color_manual(values = my_color) +
        guides(colour = guide_legend(override.aes = list(size = 2)))
    
    ggsave(paste0("dev_cell_type/the-", n, "-cluster.pdf"))
}
```
