==> 03_glomerular_data_GSE146912.Rmd <==
---
title: "03_glomerular_data_GSE146912"
author: "YinCY"
date: "2023-03-01"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# prepare glomerular cells
data from GSE146912
```{r}
library(magrittr)
library(dplyr)
library(data.table)
library(tibble)
library(SingleCellExperiment)

# normal
normal <- fread("../data/kidney/kidney-map/GSE146912/Normal_1.txt.gz", 
                sep = "\t", 
                header = TRUE)
gene_df <- normal[, 1:2]
gene_df <- gene_df %>% dplyr::rename(ensembl = IGIS) %>% as.data.frame()
rownames(gene_df) <- gene_df$ensembl
cell_df <- data.frame(cell_ids = colnames(normal)[-c(1,2)], 
                      location = "Cortex")
rownames(cell_df) <- cell_df$cell
normal %>% dim
normal[1:10, 1:5]
normal <- normal %>% column_to_rownames("IGIS") %>% dplyr::select(-SYMBOL)

glomer_sce <- SingleCellExperiment(assays = list(counts = as(as.matrix(normal), "CsparseMatrix")), 
                                   rowData = gene_df, 
                                   colData = cell_df)
glomer_sce
```

```{r}
# quality control
library(scran)
library(BiocParallel)
library(magrittr)
library(scDblFinder)
library(scater)
library(bluster)

is_mito <- grepl("^mt-", rowData(glomer_sce)$SYMBOL, ignore.case = F)
is_mito %>% table(useNA = "ifany")
rowData(glomer_sce)$SYMBOL[is_mito] %>% sort
colData(glomer_sce)$mito_percent <- round((colSums(assay(glomer_sce[is_mito, ], 1)) / colSums(assay(glomer_sce, 1))) * 100, digits = 3)
colData(glomer_sce)

set.seed(101)
qclusters <- quickCluster(glomer_sce, 
                          min.size = 100, 
                          method = "igraph", 
                          use.ranks = T, 
                          BPPARAM = MulticoreParam(workers = 12))
qclusters %>% table
colData(glomer_sce)$qclusters <- qclusters

qc <- perCellQCMetrics(glomer_sce, subsets = list(mito = is_mito))
qc$subsets_mito_percent %>% quantile(probs = seq(0, 1, 0.05))

qc_stat <- perCellQCFilters(x = qc, 
                            sub.fields = "subsets_mito_percent")
qc_stat %>% as.data.frame() %>% colSums()
glomer_sce <- glomer_sce[, !(qc_stat$low_n_features | qc_stat$low_lib_size | qc$subsets_mito_percent > 5)]
```

```{r}
# doublet detection
set.seed(102)
dbs <- scDblFinder(glomer_sce)
dbs$scDblFinder.class %>% table
glomer_sce <- glomer_sce[, dbs$scDblFinder.class == "singlet"]
glomer_sce
```


```{r}
# normalization
set.seed(103)
libsize <- calculateSumFactors(assay(glomer_sce, 1), clusters = glomer_sce$qclusters)
libsize %>% range
sizeFactors(glomer_sce) <- libsize

glomer_sce <- logNormCounts(glomer_sce)
glomer_sce
```


```{r}
# model gene variance
glomer_variance <- modelGeneVar(glomer_sce)
glomer_variance[order(glomer_variance$bio, decreasing = T), ]

# feature selection
hvgs <- getTopHVGs(glomer_variance, fdr.threshold = 0.001)
hvgs %>% str
```


```{r}
# dimensionality reduction
set.seed(104)
glomer_sce <- runPCA(glomer_sce, subset_row = hvgs)

library(PCAtools)
elbow <- findElbowPoint(attr(reducedDim(glomer_sce, "PCA"), "percentVar"))
data.frame(PC = 1:length(attr(reducedDim(glomer_sce, "PCA"), "percentVar")), 
           percentvar = attr(reducedDim(glomer_sce, "PCA"), "percentVar")) %>% 
    ggplot(aes(PC, attr(reducedDim(glomer_sce, "PCA"), "percentVar"))) +
    geom_point() +
    geom_line() +
    geom_vline(xintercept = elbow, linetype = 2)

reducedDim(glomer_sce, "PCA") <- reducedDim(glomer_sce, "PCA")[, 1:7]

set.seed(105)
glomer_sce <- runTSNE(glomer_sce, use.dimred = "PCA", perplexity = 10)

plotReducedDim(glomer_sce, 
               dimred = "TSNE", 
               colour_by = "qclusters", 
               point_size = 0.5)
```


```{r}
# clustering
set.seed(106)
clusters <- clusterRows(reducedDim(glomer_sce, "PCA"), 
                        BLUSPARAM = SNNGraphParam(cluster.fun = "louvain", 
                                                  k = 10,
                                                 BPPARAM = MulticoreParam(10)))
clusters %>% table

colLabels(glomer_sce) <- factor(clusters)
```


```{r}
source("functions/idconv.R")
library(patchwork)
plotReducedDim(glomer_sce, 
               dimred = "TSNE", 
               colour_by = idconv("Cdh5"),
               # colour_by = "label",
               text_by = "label", 
               point_size = 0.4)
```

```{r}
# cell type annotation
cell_type <- c("1" = "Mesangial Cells A", 
               "2" = "Parietal Epithelial", 
               "3" = "Endothelial",
               "4" = "Mesangial Cells B",
               "5" = "Podocyte", 
               "6" = "Podocyte",
               "7" = "Glomerular Endothelial", 
               "8" = "Podocyte", 
               "9" = "Immune Cells", 
               "10" = "Glomerular Endothelial", 
               "11" = "Glomerular Endothelial",
               "12" = "Glomerular Endothelial", 
               "13" = "VSMC"
               )

glomer_sce$cell_type <- cell_type[glomer_sce$label]
glomer_sce %>% colData %>% .$cell_type %>% table()

# remove immune cells
glomer_sce <- glomer_sce[, !(glomer_sce$cell_type == "Immune Cells")]
glomer_sce %>% colData
```


```{r}
plotTSNE(glomer_sce, 
         colour_by = "cell_type", 
         text_by = "cell_type", 
         point_size = 0.5, 
         text_size = 3, 
         force = 2) +
    theme(legend.position = "none")

saveRDS(glomer_sce, "../data/kidney/kidney-map/GSE146912/glomer_sce.rds")
```
