==> 05_integration.Rmd <==
---
title: "integration"
author: "YinCY"
date: 2023-0330
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Integration with harmony
```{r}
library(dplyr)
library(scran)
library(scater)
library(batchelor)
library(harmony)

sym_sce <- readRDS("../data/kidney/kidney-map/GSE78845/sym_sce.rds")
drg_sce <- readRDS("../data/kidney/kidney-map/GSE175421/drg_sce.rds")
glomer_sce <- readRDS("../data/kidney/kidney-map/GSE146912/glomer_sce.rds")
dev_cell_sce <- readRDS("../data/kidney/kidney-map/GSE129798/dev_cell_sce.rds")


sample_meta <- data.frame(
    cell_ids = c(colnames(drg_sce), colnames(sym_sce), 
                 colnames(glomer_sce), colnames(dev_cell_sce)),
    
    geo_accession = rep(c("GSE175421", "GSE78845", "GSE146912", "GSE129798"), 
                        times = c(dim(drg_sce)[2], dim(sym_sce)[2], 
                                  dim(glomer_sce)[2], dim(dev_cell_sce)[2])), 
    
    cell_type = c(drg_sce$cell_type, sym_sce$cell_type, 
                  glomer_sce$cell_type, dev_cell_sce$cell_type), 
    
    mito_percent = c(drg_sce$mito_percent, sym_sce$mito_percent, 
                                glomer_sce$mito_percent, dev_cell_sce$mito_percent), 
    
    location = factor(c(drg_sce$location, sym_sce$location, glomer_sce$location, dev_cell_sce$location), 
                      levels = c("Cortex", "outer medulla", "inner medulla", "paravertebral")))
rownames(sample_meta) <- sample_meta$cell_ids
sample_meta

source("functions/harmoniseSCE.R")
counts <- harmonizeSCE(sce_list = list(drg_sce, sym_sce, glomer_sce, dev_cell_sce), method = "harmony")

library(AnnotationHub)
ah <- AnnotationHub()
query(ah, c("Mus musculus", "EnsDb"))
ensdb <- ah[["AH109655"]]
gene_df <- data.frame(ensembl = rownames(counts), 
                      symbol = mapIds(ensdb, keys = rownames(counts), keytype = "GENEID", column = "SYMBOL"))
gene_df %>% head

sces <- SingleCellExperiment(assays = list(counts = counts), 
                             rowData = gene_df, 
                             colData = sample_meta)
sces
```



```{r}
set.seed(101)
sces <- multiBatchNorm(sces, batch = sces$geo_accession, normalize.all = FALSE)

vars <- modelGeneVar(sces, block = sces$geo_accession)
hvgs <- getTopHVGs(vars, n = 5000)
hvgs %>% str

set.seed(102)
npcs <- calculatePCA(sces, subset_row = hvgs)
library(PCAtools)
elbow <- findElbowPoint(attr(npcs, "percentVar"))
reducedDim(sces, "PCA") <- npcs[, 1:50]


set.seed(103)
sces <- RunHarmony(object = sces, 
                   group.by.vars = "geo_accession", 
                   reduction.save = "harmony", 
                   epsilon.harmony = 1e-06)

set.seed(104)
sces <- runTSNE(sces, dimred = "harmony")



library(viridis)
plotTSNE(sces, 
         colour_by = "cell_type", 
         point_size = 0.05, 
         text_by = "cell_type", 
         text_size = 1.5) +
    scale_color_viridis(discrete = TRUE) +
    theme(legend.position = "none") +
    guides(color = guide_legend(override.aes = list(size = 3)))

saveRDS(sces, "../data/kidney/kidney-map/integrated_data/harmony_integrated.rds")
```


```{r}
library(harmony)
library(patchwork)
library(ggplot2)
library(scater)

sces <- readRDS("../data/kidney/kidney-map/integrated_data/harmony_integrated.rds")
sces$location <- factor(sces$location, levels = c("Cortex", "outer medulla", "inner medulla", "paravertebral"))

p_harmony_integrated_01 <- plotReducedDim(sces, 
               dimred = "TSNE", 
               color_by = "cell_type", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    theme(legend.position = "none") 
saveRDS(p_harmony_integrated_01, 
        "../data/kidney/kidney-map/integrated_data/p_harmony_integrated_01.rds")

p_harmony_integrated_02 <- plotReducedDim(sces, 
               dimred = "TSNE", 
               color_by = "geo_accession", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    theme(legend.position = "right") +
    scale_color_discrete(name = "dataset") +
    guides(color = guide_legend(override.aes = list(size = 2))) 
saveRDS(p_harmony_integrated_02, 
        "../data/kidney/kidney-map/integrated_data/p_harmony_integrated_02.rds")

p_harmony_integrated_03 <- plotReducedDim(sces, 
               dimred = "TSNE", 
               color_by = "location", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    theme(legend.position = "right") +
    scale_color_discrete(name = "region") +
    guides(color = guide_legend(override.aes = list(size = 2))) +
    plot_layout(ncol = 1)

saveRDS(p_harmony_integrated_03, 
        "../data/kidney/kidney-map/integrated_data/p_harmony_integrated_03.rds")
```



# Integration with RISC
```{r}
library(SingleCellExperiment)
library(RISC)
library(magrittr)

sym_sce <- readRDS("../data/kidney/kidney-map/GSE78845/sym_sce.rds")
drg_sce <- readRDS("../data/kidney/kidney-map/GSE175421/drg_sce.rds")
glomer_sce <- readRDS("../data/kidney/kidney-map/GSE146912/glomer_sce.rds")
dev_cell_sce <- readRDS("../data/kidney/kidney-map/GSE129798/dev_cell_sce.rds")

sym_sce$geo_accession <- "GSE78845"
drg_sce$geo_accession <- "GSE175421"
glomer_sce$geo_accession <- "GSE146912"
dev_cell_sce$geo_accession <- "GSE129798"

sces <- list(dev_cell = dev_cell_sce, drg = drg_sce, glomer = glomer_sce, sym = sym_sce)
source("functions/harmoniseSCE.R")
sces <- harmonizeSCE(sces, method = "RISC")

# library(batchelor)
# set.seed(101)
# sces <- multiBatchNorm(sces[[1]], sces[[2]], sces[[3]], sces[[4]])

source("functions/createRISC.R")
sces <- createRISC(sce_list = sces)

rm(dev_cell_sce, drg_sce, sym_sce, glomer_sce)
gc()

pro <- function(x, ...){
        # do not filtering, is been done
        x <- scFilter(x,min.UMI = 0, max.UMI = Inf, min.gene = 0, 
                      min.cell = 0, mitochon = 1, is.filter = FALSE)
        
        # normalize the raw counts
        x <- scNormalize(x, ncore = 3)
        
        # find high variable genes
        x <- scDisperse(x)
    return(x)
}


set.seed(101)
sces <- lapply(sces, pro)
gc()

var_genes <- Reduce(union, lapply(sces, function(x){x@vargene})) %>% unique()
var_genes %>% str

sces <- scMultiIntegrate(objects = sces, 
                         eigens = 30, 
                         var.gene = var_genes, 
                         align = "Predict", 
                         npc = 50, 
                         ncore = 4, 
                         seed = 102)
gc()

set.seed(103)
sces <- scTSNE(sces, npc = 50, use = "PLS", perplexity = 30)

saveRDS(sces, "../data/kidney/kidney-map/integrated_data/RISC_integrated.rds")
```

```{r}
library(RISC)
library(dplyr)
library(tidyr)
library(ggplot2)
library(SingleCellExperiment)
library(patchwork)

sces <- readRDS("../data/kidney/kidney-map/integrated_data/RISC_integrated.rds")
df1 <- sces@coldata %>% as.data.frame()
df2 <- sces@DimReduction$cell.tsne %>% as.data.frame() %>% tibble::rownames_to_column("Barcode")
df <- left_join(df2, df1)
df %>% dim
df$Set %>% table
df$Set <- sapply(df$Set, function(x){switch(as.character(x), 
                                            "Set1" = "GSE129798", 
                                            "Set2" = "GSE175421", 
                                            "Set3" = "GSE146912", 
                                            "Set4" = "GSE78845")})

cell_coord <- df %>% group_by(cell_type) %>% summarise(x = median(tSNE1), y = median(tSNE2))
df$location <- factor(df$location, levels = c("Cortex", "outer medulla", "inner medulla", "paravertebral"))


p_RISC_integrated_01 <- df %>% 
    ggplot(aes(tSNE1, tSNE2)) +
    geom_point(aes(color = cell_type), size = 0.2) +
    geom_text(data = cell_coord, aes(x, y, label = cell_type), size = 2) +
    theme_classic() +
    theme(legend.position = "none")
saveRDS(p_RISC_integrated_01, 
        "../data/kidney/kidney-map/integrated_data/p_RISC_integrated_01.rds")

p_RISC_integrated_02 <- df %>% 
    ggplot(aes(tSNE1, tSNE2)) +
    geom_point(aes(color = Set), size = 0.2) +
    geom_text(data = cell_coord, aes(x, y, label = cell_type), size = 2) +
    theme(legend.position = "right") +
    scale_color_discrete(name = "dataset") +
    theme_classic() +
    guides(color = guide_legend(override.aes = list(size = 2))) 
saveRDS(p_RISC_integrated_02, 
        "../data/kidney/kidney-map/integrated_data/p_RISC_integrated_02.rds")
    
p_RISC_integrated_03 <- df %>% 
    ggplot(aes(tSNE1, tSNE2)) +
    geom_point(aes(color = location), size = 0.2) +
    geom_text(data = cell_coord, aes(x, y, label = cell_type), size = 2) +
    scale_color_discrete(name = "region") +
    theme(legend.position = "right") +
    theme_classic() +
    guides(color = guide_legend(override.aes = list(size = 2))) +
    plot_layout(ncol = 1)

saveRDS(p_RISC_integrated_03, 
        "../data/kidney/kidney-map/integrated_data/p_RISC_integrated_03.rds")
```



# Integration with fastMNN
```{r}
library(SingleCellExperiment)
library(scran)
library(batchelor)
library(scater)
library(dplyr)
library(bluster)


sym_sce <- readRDS("../data/kidney/kidney-map/GSE78845/sym_sce.rds")
drg_sce <- readRDS("../data/kidney/kidney-map/GSE175421/drg_sce.rds")
glomer_sce <- readRDS("../data/kidney/kidney-map/GSE146912/glomer_sce.rds")
dev_cell_sce <- readRDS("../data/kidney/kidney-map/GSE129798/dev_cell_sce.rds")

sces <- list(dev_cell = dev_cell_sce, 
             glomer = glomer_sce, 
             drg = drg_sce, 
             sym = sym_sce)

source("functions/harmoniseSCE.R")
sces <- harmonizeSCE(sces, method = "fastMNN")

set.seed(101)
sces <- multiBatchNorm(sces[[1]], sces[[2]], sces[[3]], sces[[4]])

set.seed(102)
dec_list <- lapply(sces, modelGeneVar)
combined_dec <- combineVar(dec_list[[1]], dec_list[[2]], dec_list[[3]], dec_list[[4]])
hvgs <- getTopHVGs(combined_dec, n = 3000)
hvgs %>% str

# for findMarkers
hvgs1 <- getTopHVGs(combined_dec, n = 5000)
saveRDS(hvgs1, "../data/kidney/kidney-map/integrated_data/hvgs.rds")


rm(dev_cell_sce, glomer_sce, drg_sce, sym_sce)
gc()
```


```{r}
set.seed(103)
combined_sces <- correctExperiments(sces[[1]], sces[[2]], sces[[3]], sces[[4]], 
                                    include.rowdata = TRUE, 
                                    PARAM = FastMnnParam())

set.seed(104)
combined_sces <- runTSNE(combined_sces, 
                         dimred = "corrected", 
                         perplexity = 30)
# set.seed(110)
# combined_sces <- runUMAP(combined_sces, dimred = "corrected")
```


## clustering
```{r}
set.seed(105)
clusters <- clusterRows(x = reducedDim(combined_sces, "TSNE"), 
                        BLUSPARAM = NNGraphParam(k = 20, 
                                                 cluster.fun = "walktrap", 
                                                 cluster.args = list(steps = 5)))
clusters %>% table
combined_sces$clusters <- factor(clusters)

saveRDS(combined_sces, "../data/kidney/kidney-map/integrated_data/fastMNN_integrated_after_clustering.rds")


plotTSNE(combined_sces, 
         colour_by = "clusters", 
         text_by = "clusters", 
         text_size = 2,
         point_size = 0.2) +
    theme(legend.position = "right") +
    guides(color = guide_legend(override.aes = list(size = 2)))
ggsave("../data/kidney/kidney-map/integrated_data/clusters.tiff", width = 10)

plotTSNE(combined_sces, 
         colour_by = "clusters", 
         text_by = "clusters", 
         text_size = 2,
         point_size = 0.2) +
    theme(legend.position = "none", 
          axis.line = element_blank(), 
          axis.title = element_blank(), 
          axis.text = element_blank(), 
          axis.ticks = element_blank()) 

ggsave("../data/kidney/kidney-map/integrated_data/clusters_for_subcluster.tiff", width = 7)
```


```{r}
for(n in unique(colData(combined_sces)$clusters)){
    my_color <- rep("grey", length(colData(combined_sces)$clusters))
    names(my_color) <- as.character(colData(combined_sces)$clusters)
    my_color[colData(combined_sces)$clusters == n] <- "red"
    
    plotReducedDim(object = combined_sces, 
                   dimred = "TSNE", 
                   colour_by = "clusters", 
                   text_by = "clusters", 
                   point_size = 0.2, 
                   text_size = 2.5) +
        scale_color_manual(values = my_color) +
        guides(colour = guide_legend(override.aes = list(size = 2)))
    
    ggsave(paste0("sces/the-", n, "-cluster.pdf"))
}


for(n in c("LOH_MTAL", "LOH_CTAL")){
    my_color <- rep("grey", length(colData(sces)$cell_type))
    names(my_color) <- as.character(colData(sces)$cell_type)
    my_color[colData(sces)$cell_type == n] <- "red"
    
    plotReducedDim(object = sces, 
                   dimred = "TSNE", 
                   colour_by = "cell_type", 
                   text_by = "cell_type", 
                   point_size = 0.2, 
                   text_size = 2.5) +
        scale_color_manual(values = my_color) +
        guides(colour = guide_legend(override.aes = list(size = 2)))
    
    ggsave(paste0("sces/the-", n, "-cluster.pdf"))
}
table(combined_sces$cell_type, combined_sces$clusters) %>% as.matrix()  %>% t
```



## annotation cell types
```{r}
library(SingleR)
library(magrittr)

ref_mouse_tubule_sce <- readRDS("../data/kidney/kidney-map/GSE150338/ref_mouse_tubule_sce.rds")
agg_combined_sces <- aggregateAcrossCells(combined_sces, 
                                          ids = combined_sces$clusters, 
                                          statistics = "sum", 
                                          use.assay.type = "logcounts")

set.seed(109)
tubule_pred <- SingleR(test = agg_combined_sces, 
                       ref = ref_mouse_tubule_sce, 
                       labels = ref_mouse_tubule_sce$cell_type, 
                       assay.type.test = "logcounts", 
                       assay.type.ref = "tpm", 
                       de.method = "classic")

tubule_pred$labels %>% table
cs <- rownames(tubule_pred)
```

```{r}
tiff("../data/kidney/kidney-map/results/Supplemental/tubule_pre.tiff", 
     width = 2000, height = 2000)
tubule_pred$scores %>% as.data.frame(row.names = cs) %>% as.matrix() %>% 
    pheatmap::pheatmap(angle_col = 315, 
                       cluster_cols = T, 
                       show_rownames = T,
                       show_colnames = T,
                       color = viridis::viridis(n = 1000, option = "H"), 
                       fontsize_row = 5, 
                       display_numbers = TRUE, 
                       fontsize_number = 4, 
                       number_format = "%.3f")
dev.off()


similarity_score <- tubule_pred$scores %>% as.data.frame(row.names = cs)

library(ggheatmap)
ggheatmap(data = similarity_score, 
          legendName = "Similarity \nScore", 
          cluster_rows = T, 
          cluster_cols = T, 
          hclust_method = "ward.D2", 
          color = colorRampPalette(colors = rev(RColorBrewer::brewer.pal(7, "RdYlBu")))(100), 
          tree_height_rows = 0.3) 
ggsave("../data/kidney/kidney-map/results/Supplemental/tubule_pre.tiff", 
       height = 12, width = 8)
```


## sub clusters
```{r}
# 39, 48:  NK, Cd4 T, Cd8 T 

c39_48 <- combined_sces[, combined_sces$clusters %in% c(39, 48)]
source("functions/idconv.R")
hvgs <- c("ENSMUSG00000032094", "ENSMUSG00000002033", 
          "ENSMUSG00000030579", "ENSMUSG00000023274", 
          "ENSMUSG00000053977", "ENSMUSG00000053044")


set.seed(101)
c39_48<- runPCA(c39_48, 
              subset_row = hvgs, 
              ncomponents = 4)

set.seed(102)
c39_48 <- runTSNE(c39_48, dimred = "PCA", preplexity = 30)

set.seed(103)
c39_48_clusters <- clusterRows(reducedDim(c39_48, "PCA"), 
                            BLUSPARAM = NNGraphParam(k = 50))
c39_48_clusters %>% table
colLabels(c39_48) <- c39_48_clusters

colLabels(c39_48) <- sapply(c39_48_clusters, FUN = function(x){
  switch(as.character(x), 
         "1" = "2",
         "2" = "2", 
         "3" = "1", 
         "4" = "2", 
         "5" = "3", 
         "6" = "2")
})
# 1 NK: 3
# 2 Cd4 T Cells: 1, 2, 4, 6
# 3 Cd8 T Cells: 5
c39_48$cell_type <- c("1" = "NK Cells", "2" = "Cd4 T Cells", "3" = "Cd8 T Cells")[c39_48$label]

colLabels(c39_48) <- paste("39_48.",colLabels(c39_48), sep = "") 
labels <- as.character(combined_sces$clusters)
labels[labels %in% c(39, 48)] <- colLabels(c39_48)
combined_sces$clusters <- labels

saveRDS(c39_48, "../data/kidney/kidney-map/integrated_data/sub-clusters/c39_48.rds")

source("functions/idconv.R")
library(patchwork)
plotTSNE(c39_48, 
         colour_by = "cell_type",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_manual(name = "cell type", values = c("green", "blue", "black")) +
plotTSNE(c39_48, 
         colour_by = idconv("Cd4"),
         text_by = "cell_type",
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Cd4", low = "grey80", high = "red") +
plotTSNE(c39_48, 
         # colour_by = "label",
         colour_by = idconv("Cd8a"),
         text_by = "cell_type",
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Cd8a", low = "grey80", high = "red") +
plotTSNE(c39_48, 
         # colour_by = "label",
         colour_by = idconv("Gzma"),
         text_by = "cell_type",
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Gzma", low = "grey80", high = "red")

ggsave("../data/kidney/kidney-map/integrated_data/sub-clusters/T_NK_cell_clusters.tiff")
```


```{r}
# CD_IC: CCD_IC, OMCD_IC, IMCD_IC
library(readxl)

c21_41 <- combined_sces[, combined_sces$clusters %in% c(21, 41)]
ccd <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                  sheet = "CCD_797") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)
omcd <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                  sheet = "OMCD_757") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)
imcd <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                  sheet = "IMCD_1951") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)

hvgs <- unique(c(ccd, omcd, imcd))
hvgs <- hvgs[hvgs %in% rownames(c21_41)]

set.seed(101)
c21_41 <- runPCA(c21_41, 
              subset_row = hvgs, 
              ncomponents = 4)

set.seed(102)
c21_41 <- runTSNE(c21_41, dimred = "PCA", preplexity = 30)

set.seed(103)
c21_41_clusters <- clusterRows(reducedDim(c21_41, "PCA"), 
                            BLUSPARAM = NNGraphParam(k = 50))
c21_41_clusters %>% table
colLabels(c21_41) <- c21_41_clusters


# 1:IMCD_IC, 2: CCD_IC, 3: OMCD_IC
colLabels(c21_41) <- paste("21_41.",colLabels(c21_41), sep = "")
c21_41$cell_type <- c("21_41.1" = "IMCD_IC", "21_41.2" = "CCD_IC", "21_41.3" = "OMCD_IC")[c21_41$label]
c21_41$cell_type %>% table
saveRDS(c21_41, "../data/kidney/kidney-map/integrated_data/sub-clusters/CD_IC.rds")

labels <- as.character(combined_sces$clusters)
labels[labels %in% c(21, 41)] <- colLabels(c21_41)
combined_sces$clusters <- labels

library(patchwork)
plotTSNE(c21_41, 
         colour_by = "cell_type",
         # colour_by = "",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_manual(name = "cell type", values = c("green", "blue", "black")) +

plotTSNE(c21_41, 
         colour_by = idconv("Slc26a4"),
         # colour_by = "",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Slc26a4", low = "grey80", high = "red") +

plotTSNE(c21_41, 
         colour_by = idconv("Aqp6"),
         # colour_by = "label",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Aqp6", low = "grey80", high = "red") +

plotTSNE(c21_41, 
         colour_by = idconv("Aqp6"),
         # colour_by = "label",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Kit", low = "grey80", high = "red") 

ggsave("../data/kidney/kidney-map/integrated_data/sub-clusters/CD_IC_clusters.tiff")
```


```{r}
# LOH: LOH_DTL2, LOH_DTL3
cs <- combined_sces[, combined_sces$clusters == 7]
dtl2 <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                   sheet = "DTL2_682") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)
dtl3 <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                   sheet = "DTL3_282") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)

hvgs <- unique(c(dtl2, dtl3))
hvgs <- hvgs[hvgs %in% rownames(cs)]

set.seed(101)
cs <- runPCA(cs, 
              subset_row = hvgs, 
              ncomponents = 8)

set.seed(102)
cs <- runTSNE(cs, dimred = "PCA", preplexity = 40)

set.seed(103)
cs_clusters <- clusterRows(reducedDim(cs, "PCA"), 
                            BLUSPARAM = NNGraphParam(k = 20))
cs_clusters %>% table
colLabels(cs) <- cs_clusters

# 1 DTL3: 1, 2
# 2 DTL2: 3, 4 
cs$cell_type <- c("1" = "LOH_DTL3", "2" = "LOH_DTL3", "3" = "LOH_DTL2", "4" = "LOH_DTL2")[cs$label]

colLabels(cs) <- sapply(cs_clusters, FUN = function(x){
  switch(as.character(x), 
         "1" = "1",
         "2" = "1", 
         "3" = "2", 
         "4" = "2")
})

colLabels(cs) <- paste("7.", colLabels(cs), sep = "") 
saveRDS(cs, "../data/kidney/kidney-map/integrated_data/sub-clusters/LOH_DTL23.rds")

labels <- as.character(combined_sces$clusters)
labels[labels == 7] <- colLabels(cs)
combined_sces$clusters <- labels

plotTSNE(cs, 
         colour_by = "cell_type",
         # colour_by = "",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_manual(name = "cell type", values = c("green", "blue")) +

plotTSNE(cs, 
         colour_by = idconv("Fcgbp"),
         # colour_by = "",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) + 
    scale_color_gradient(name = "Fcgbp", low = "grey80", high = "red") +

plotTSNE(cs, 
         colour_by = idconv("Slc39a8"),
         # colour_by = "",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Slc39a8", low = "grey80", high = "red") +
    plot_layout(nrow = 2)

ggsave("../data/kidney/kidney-map/integrated_data/sub-clusters/LOH_DTL2-3.tiff")
```


```{r}
# 24, 56, 61
# LOH: LOH_MTAL, LOH_CTAL
cs <- combined_sces[, combined_sces$clusters %in% c(24, 56, 61)]
mtal <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                   sheet = "MTAL_1020") %>% 
    dplyr::filter(FDR < 0.1) %>% 
    dplyr::pull("Gene_id")
ctal <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                   sheet = "CTAL_1092") %>% 
    dplyr::filter(FDR < 0.1) %>% 
    dplyr::pull(Gene_id)

hvgs <- unique(c(mtal, ctal))
hvgs <- hvgs[hvgs %in% rownames(cs)]

set.seed(101)
cs <- runPCA(cs, 
              subset_row = hvgs, 
              ncomponents = 20)

set.seed(102)
cs <- runTSNE(cs, dimred = "PCA", preplexity = 20)

set.seed(103)
cs_clusters <- clusterRows(reducedDim(cs, "PCA"), 
                            BLUSPARAM = NNGraphParam(k = 20))
cs_clusters %>% table
colLabels(cs) <- cs_clusters

# 1 MTAL: 2,4
# 2 CTAL: 1,3
cs$cell_type <- c("1" = "LOH_CTAL", "2" = "LOH_MTAL", "3" = "LOH_CTAL", "4" = "LOH_MTAL")[cs$label]
colLabels(cs) <- sapply(cs_clusters, FUN = function(x){
  switch(as.character(x), 
         "1" = "2",
         "2" = "1", 
         "3" = "2", 
         "4" = "1")
})

colLabels(cs) <- paste("24_61.",colLabels(cs), sep = "") 
saveRDS(cs, "../data/kidney/kidney-map/integrated_data/sub-clusters/LOH_CTAL_MTAL.rds")

labels <- as.character(combined_sces$clusters)
labels[labels %in% c(24, 56, 61)] <- colLabels(cs)
combined_sces$clusters <- labels

library(patchwork)
plotTSNE(cs, 
         colour_by = "cell_type",
         # colour_by = "",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_manual(name = "cell type", values = c("green", "blue")) +

plotTSNE(cs, 
         colour_by = idconv("Nccrp1"),
         # colour_by = "",
         text_by = "cell_type",
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Nccrp1", low = "grey80", high = "red") +

plotTSNE(cs, 
         colour_by = idconv("Pth1r"),
         # colour_by = "",
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Pth1r", low = "grey80", high = "red") +
    plot_layout(nrow = 2)

ggsave("../data/kidney/kidney-map/integrated_data/sub-clusters/LOH_MTAL_CTAL.tiff")
```


```{r, message=FALSE}
# CCD_PC, OMCD_PC, IMCD_PC: 12, 15, 36, 49, 59
cs <- combined_sces[, combined_sces$clusters %in% c(12, 15, 36, 49, 59)]

ccd <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                  sheet = "CCD_797") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)

omcd <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                   sheet = "OMCD_757") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)

imcd <- read_excel("../data/kidney/kidney-map/GSE150338/supplementary_files/ASN.2020101406SupplementaryData4.xlsx", 
                   sheet = "IMCD_1951") %>% 
    dplyr::filter(FDR < 0.01) %>% 
    dplyr::pull(Gene_id)

hvgs <- unique(c(ccd, omcd, imcd))
hvgs <- hvgs[hvgs %in% rownames(cs)]

set.seed(101)
cs <- runPCA(cs, 
              subset_row = hvgs, 
              ncomponents = 10)

set.seed(102)
cs <- runTSNE(cs, dimred = "PCA", preplexity = 30)

set.seed(103)
cs_clusters <- clusterRows(reducedDim(cs, "PCA"), 
                            BLUSPARAM = NNGraphParam(k = 20, cluster.fun = "louvain"))
cs_clusters %>% table
colLabels(cs) <- cs_clusters
saveRDS(cs, "../data/kidney/kidney-map/integrated_data/sub-clusters/CD_PC.rds")

# 1 CCD_PC: 1, 2 : Calb1, Tep1
# 2 OMCD_PC: 3, 4, 5 : Btc, Itm2a, Slc4a5, 
# 3 IMCD_PC: 6, 7: Aldh1a3

cs$cell_type <- c("1" = "CCD_PC", "2" = "CCD_PC", "3" = "OMCD_PC", 
                  "4" = "OMCD_PC", "5" = "OMCD_PC", "6" = "IMCD_PC", 
                  "7" = "IMCD_PC")[cs$label]

colLabels(cs) <- sapply(cs_clusters, FUN = function(x){
  switch(as.character(x), 
         "1" = "1",
         "2" = "1", 
         "3" = "2", 
         "4" = "2", 
         "5" = "2", 
         "6" = "3", 
         "7" = "3")
})

colLabels(cs) <- paste("12_59.",colLabels(cs), sep = "") 

labels <- as.character(combined_sces$clusters)
labels[labels %in% c(12, 15, 36, 49, 59)] <- colLabels(cs)
combined_sces$clusters <- labels


library(patchwork)
plotTSNE(cs, 
         colour_by = "cell_type", 
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_manual(name = "cell type", values = c("green", "blue", "black")) +

plotTSNE(cs, 
         colour_by = idconv("Calb1"), 
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Calb1", low = "grey80", high = "red") +

plotTSNE(cs, 
         colour_by = idconv("Btc"), 
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Btc", low = "grey80", high = "red") +

plotTSNE(cs, 
         colour_by = idconv("Aldh1a3"), 
         text_by = "cell_type", 
         text_size = 4, 
         point_size = 1) +
    scale_color_gradient(name = "Aldh1a3", low = "grey80", high = "red") 

ggsave("../data/kidney/kidney-map/integrated_data/sub-clusters/CD_PC_clusters.tiff")
```


```{r}
cell_type_annotation <- c(
    "1" = "LOH_ATL",
    "2" = "Schwann Cells",
    "3" = "Satellite Glia A",
    "4" = "Glomerular Endothelial",
    "5" = "Sensory Neurons",
    "6" = "CNT",
    "7.1" = "LOH_DTL3",
    "7.2" = "LOH_DTL2",
    "8" = "Macrophages",
    "9" = "Mesangial Cells A",
    "10" = "Dendritic Cells",
    "11" = "Endothelial",
    "12_59.1" = "CCD_PC",
    "12_59.2" = "OMCD_PC",
    "12_59.3" = "IMCD_PC",
    "13" = "PTS1",
    "14" = "Sympathetic Neurons",
    "16" = "Podocytes",
    "17" = "Satellite Glia B",
    "18" = "Macrophages",
    "19" = "paravertebral Endothelial",
    "20" = "PTS3",
    "21_41.1" = "IMCD_IC",
    "21_41.2" = "CCD_IC",
    "21_41.3" = "OMCD_IC",
    "22" = "Endothelial",
    "23" = "Endothelial",
    "24_61.1" = "LOH_MTAL",
    "24_61.2" = "LOH_CTAL",
    "25" = "Parietal Epithelial",
    "26" = "doublets",
    "27" = "Endothelial",
    "28" = "PTS1",
    "29" = "Endothelial",
    "30" = "DCT",
    "31" = "B Cells",
    "32" = "Mesangial Cells B",
    "33" = "PTS2",
    "34" = "Schwann Cells",
    "35" = "Podocytes",
    "37" = "PTS1",
    "38" = "Endothelial",
    "39_48.1" = "NK Cells",
    "39_48.2" = "Cd4 T Cells",
    "39_48.3" = "Cd8 T Cells",
    "40" = "Pericytes",
    "42" = "Monocytes",
    "43" = "LOH_ATL",
    "44" = "DCT",
    "45" = "Satellite Glia A",
    "46" = "VSMC",
    "47" = "Fibroblast",
    "50" = "Endothelial",
    "51" = "Sensory Neurons",
    "52" = "LOH_ATL",
    "53" = "Monocytes",
    "54" = "doublets",
    "55" = "Podocytes",
    "57" = "doublets",
    "58" = "doublets",
    "60" = "VSMC",
    "62" = "Satellite Glia B",
    "63" = "doublets",
    "64" = "PTS2",
    "65" = "LOH_DTL1",
    "66" = "Fibroblast",
    "67" = "doublets",
    "68" = "doublets",
    "69" = "Endothelial",
    "70" = "doublets",
    "71" = "doublets",
    "72" = "Endothelial",
    "73" = "doublets",
    "74" = "doublets"
)



cell_type_annotation
combined_sces$cell_type <- cell_type_annotation[combined_sces$clusters]

# convert some sympathetic neurons to sensory neurons, 
# those sensory neurons clustered with sympathetic neurons at cluster 14, but it's from DRG ganglia

colData(combined_sces) <- transform(colData(combined_sces), 
          cell_type = ifelse(cell_type == "Sympathetic Neurons" & geo_accession == "GSE175421", "Sensory Neurons", cell_type)) 

combined_sces$cell_type %>% table(useNA = "ifany")
combined_sces <- combined_sces[, combined_sces$cell_type != "doublets"]

plotTSNE(sces, 
         colour_by = "cell_type", 
         text_by = "cell_type", 
         point_size = 0.2, 
         text_size = 2) +
    scale_x_continuous(name = NULL) +
    scale_y_continuous(name = NULL) +
    theme(legend.position = "none", 
          axis.line = element_blank(), 
          axis.ticks = element_blank(), 
          axis.text = element_blank()) 
ggsave("../data/kidney/kidney-map/integrated_data/fastMNN_cell_type_TSNE.tiff")

saveRDS(combined_sces, "../data/kidney/kidney-map/integrated_data/fastMNN_integrated.rds")
```


```{r}
library(SingleCellExperiment)
library(scater)
library(scran)
library(dplyr)
library(ggplot2)
library(patchwork)

sces <- readRDS("../data/kidney/kidney-map/integrated_data/fastMNN_integrated.rds")
sces$location <- factor(sces$location, levels = c("Cortex", "outer medulla", "inner medulla", "paravertebral"))

p_fastMNN_integrated_01 <- plotReducedDim(sces, 
               dimred = "TSNE", 
               color_by = "cell_type", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    theme(legend.position = "none") 
saveRDS(p_fastMNN_integrated_01, 
        "../data/kidney/kidney-map/integrated_data/p_fastMNN_integrated_01.rds")

p_fastMNN_integrated_02 <- plotReducedDim(sces, 
               dimred = "TSNE", 
               color_by = "geo_accession", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    scale_color_discrete(name = "dataset") +
    theme(legend.position = "right") +
    guides(color = guide_legend(override.aes = list(size = 2)))
saveRDS(p_fastMNN_integrated_02, 
        "../data/kidney/kidney-map/integrated_data/p_fastMNN_integrated_02.rds")


p_fastMNN_integrated_03 <- plotReducedDim(sces, 
               dimred = "TSNE", 
               color_by = "location", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    scale_color_discrete(name = "region") +
    theme(legend.position = "right") +
    guides(color = guide_legend(override.aes = list(size = 2))) +
    plot_layout(ncol = 1)

saveRDS(p_fastMNN_integrated_03, 
        "../data/kidney/kidney-map/integrated_data/p_fastMNN_integrated_03.rds")
```


# Integration with Seurat
```{r}
library(Seurat)
library(magrittr)

sym_sce <- readRDS("../data/kidney/kidney-map/GSE78845/sym_sce.rds")
drg_sce <- readRDS("../data/kidney/kidney-map/GSE175421/drg_sce.rds")
glomer_sce <- readRDS("../data/kidney/kidney-map/GSE146912/glomer_sce.rds")
dev_cell_sce <- readRDS("../data/kidney/kidney-map/GSE129798/dev_cell_sce.rds")

sym_sce$geo_accession <- "GSE78845"
drg_sce$geo_accession <- "GSE175421"
glomer_sce$geo_accession <- "GSE146912"
dev_cell_sce$geo_accession <- "GSE129798"

kidney_list <- list(dev_cell = dev_cell_sce, 
                    glomer = glomer_sce, 
                    drg = drg_sce, 
                    sym = sym_sce)

# remove unused files to save space
rm(sym_sce, drg_sce, glomer_sce, dev_cell_sce)
gc()

kidney_list <- lapply(kidney_list, FUN = as.Seurat)

kidney_list <- lapply(kidney_list, 
                      FUN = SCTransform, 
                      seed.use = 101, 
                      variable.features.n = 5000, 
                      assay = "originalexp",
                      vars.to.regress = "mito_percent")

features <- SelectIntegrationFeatures(kidney_list, nfeatures = 5000)
features %>% str

kidney_list <- PrepSCTIntegration(kidney_list, anchor.features = features)
anchors <- FindIntegrationAnchors(kidney_list, normalization.method = "SCT", anchor.features = features)
rm(kidney_list);gc()

kidney_integrated <- IntegrateData(anchorset = anchors, normalization.method = "SCT")
```


```{r}
rm(anchors);gc()

set.seed(102)
kidney_integrated <- RunPCA(kidney_integrated, verbose = TRUE) %>% 
    RunTSNE(reduction = "pca", dims = 1:50)

saveRDS(kidney_integrated, "../data/kidney/kidney-map/integrated_data/Seurat_integrated.rds")
```


```{r}
library(Seurat)
library(scater)
library(scran)
library(patchwork)
library(magrittr)
library(ggplot2)

sces <- readRDS("../data/kidney/kidney-map/integrated_data/Seurat_integrated.rds")
sces <- as.SingleCellExperiment(sces)
sces$location <- factor(sces$location, levels = c("Cortex", "outer medulla", "inner medulla", "paravertebral"))
sces %>% colData

p_seurat_integrated_01 <- plotReducedDim(sces, 
               dimred = "TSNE",
               colour_by = "cell_type", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    theme(legend.position = "none") 
saveRDS(p_seurat_integrated_01, 
        "../data/kidney/kidney-map/integrated_data/p_seurat_integrated_01.rds")


p_seurat_integrated_02 <- plotReducedDim(sces, 
               dimred = "TSNE",
               colour_by = "geo_accession", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    scale_color_discrete(name = "dataset") +
    theme(legend.position = "right") +
    guides(color = guide_legend(override.aes = list(size = 2))) 
saveRDS(p_seurat_integrated_02, 
        "../data/kidney/kidney-map/integrated_data/p_seurat_integrated_02.rds")


p_seurat_integrated_03 <- plotReducedDim(sces, 
               dimred = "TSNE",
               colour_by = "location", 
               text_by = "cell_type", 
               text_size = 2, 
               point_size = 0.2) +
    scale_color_discrete(name = "region") +
    theme(legend.position = "right") +
    guides(color = guide_legend(override.aes = list(size = 2))) +
    plot_layout(ncol = 1)
saveRDS(p_seurat_integrated_03, 
        "../data/kidney/kidney-map/integrated_data/p_seurat_integrated_03.rds")
```


# integration maps
```{r}
library(patchwork)
library(ggplot2)

p_harmony_integrated_01 <- readRDS("../data/kidney/kidney-map/integrated_data/p_harmony_integrated_01.rds")
p_harmony_integrated_02 <- readRDS("../data/kidney/kidney-map/integrated_data/p_harmony_integrated_02.rds")
p_harmony_integrated_03 <- readRDS("../data/kidney/kidney-map/integrated_data/p_harmony_integrated_03.rds")

p_fastMNN_integrated_01 <- readRDS("../data/kidney/kidney-map/integrated_data/p_fastMNN_integrated_01.rds")
p_fastMNN_integrated_02 <- readRDS("../data/kidney/kidney-map/integrated_data/p_fastMNN_integrated_02.rds")
p_fastMNN_integrated_03 <- readRDS("../data/kidney/kidney-map/integrated_data/p_fastMNN_integrated_03.rds")

p_RISC_integrated_01 <- readRDS("../data/kidney/kidney-map/integrated_data/p_RISC_integrated_01.rds")
p_RISC_integrated_02 <- readRDS("../data/kidney/kidney-map/integrated_data/p_RISC_integrated_02.rds")
p_RISC_integrated_03 <- readRDS("../data/kidney/kidney-map/integrated_data/p_RISC_integrated_03.rds")

p_seurat_integrated_01 <- readRDS("../data/kidney/kidney-map/integrated_data/p_seurat_integrated_01.rds")
p_seurat_integrated_02 <- readRDS("../data/kidney/kidney-map/integrated_data/p_seurat_integrated_02.rds")
p_seurat_integrated_03 <- readRDS("../data/kidney/kidney-map/integrated_data/p_seurat_integrated_03.rds")

p_harmony_integrated_01 + p_harmony_integrated_02 + p_harmony_integrated_03 +
    p_seurat_integrated_01 + p_seurat_integrated_02 + p_seurat_integrated_03 +
    p_RISC_integrated_01 + p_RISC_integrated_02 + p_RISC_integrated_03 +
    p_fastMNN_integrated_01 + p_fastMNN_integrated_02 + p_fastMNN_integrated_03 +
    plot_layout(nrow = 3, ncol = 4, byrow = F)


ggsave("../data/kidney/kidney-map/integrated_data/integration_maps.tiff", 
       width = 40*0.5, height = 30*0.5)
```
